% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLST8.R
\name{getLST8}
\alias{getLST8}
\title{Retrieves OLI & TIRS raw values and calculates REFLECTANCE TOA, RADIANCE TOA and BRIGHTNESS TEMPERATURE}
\usage{
getLST8(dir, geometry)
}
\arguments{
\item{dir}{The folder or directory where the Landsat 8 GEOTIF files are stored.}

\item{geometry}{The geometry object of the geographical zone to retrieve the data from.}
}
\description{
\code{getLST8} Process Level 1 GeoTIFF Data Product of Landsat 8 and converts
to REFLECTANCE TOA, RADIANCE TOA and BRIGHTNESS TEMPERATURE.
}
\details{
The procedure considers as input (1) the folder where the GEOTIF data products are
stored taking into account that each data product must be decompress in one folder
and (2) a geometry object corresponding to the geographical zone to retrieve the data from
 (for instance a SpatialPolygonsDataFrame)
The output of this procedure is a CSV file including the spatial data points of the extent with the
raw data of the OLI (Band B8 panchromatic is not included) and TIRS bands and their corresponding reflectance
and radiance TOA values for the spectral bands and, for thermal bands, their brightness temperature.
The name of the CSV file is LST8_startDate_endDate.
This procedure has dependences on rgdal, raster and sp R packages.
The procedure does not download the data products. The data products can be downloaded from
http://earthexplorer.usgs.gov/

The structure of the CSV file is the following:

x: Coordinate x of the data point in the CRS of the data product
y: Coordinate y of the data point in the CRS of the data product
crs_proj: CRS projection string of the data product
date: Year + Day number of the year in the format YYYYddd
30_B1: OLI band 1
30_B2: OLI band 2
30_B3: OLI band 3
30_B4: OLI band 4
30_B5: OLI band 5
30_B6: OLI band 6
30_B7: OLI band 7
30_B9: OLI band 9
30_B10: TIRS band 10
30_B11: TIRS band 11
30_QA: Quality indicator of the measures
30_B1_RADIANCE: RADIANCE (TOA) for band 1
30_B1_REFLECTANCE: REFLECTANCE (TOA) for band 1
30_B2_RADIANCE: RADIANCE (TOA) for band 2
30_B2_REFLECTANCE: REFLECTANCE (TOA) for band 2
30_B3_RADIANCE: RADIANCE (TOA) for band 3
30_B3_REFLECTANCE: REFLECTANCE (TOA) for band 3
30_B4_RADIANCE: RADIANCE (TOA) for band 4
30_B4_REFLECTANCE: REFLECTANCE (TOA) for band 4
30_B5_RADIANCE: RADIANCE (TOA) for band 5
30_B5_REFLECTANCE: REFLECTANCE (TOA) for band 5
30_B6_RADIANCE: RADIANCE (TOA) for band 6
30_B6_REFLECTANCE: REFLECTANCE (TOA) for band 6
30_B7_RADIANCE: RADIANCE (TOA) for band 7
30_B7_REFLECTANCE: REFLECTANCE (TOA) for band 7
30_B9_RADIANCE: RADIANCE (TOA) for band 9
30_B9_REFLECTANCE: REFLECTANCE (TOA) for band 9
30_B10_TEMPERATURE: BRIGHTNESS TEMPERATURE (TOA) for band 11
30_B11_TEMPERATURE: BRIGHTNESS TEMPERATURE (TOA) for band 12
}
\author{
Rodolfo de Benito Arango, \email{rbarango@gmail.com}
}
\references{
\url{https://landsat.usgs.gov/Landsat8_Using_Product.php}

\url{http://landsat.usgs.gov/band_designations_landsat_satellites.php}

\url{http://earthexplorer.usgs.gov}
}
\keyword{Landsat8}
\keyword{OLI}
\keyword{TIRS}
\keyword{TOA}

